% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advance_family.R
\name{advance.family}
\alias{advance.family}
\title{Advance a family}
\usage{
advance.family(genome, pop.mat, pop.type = "random", cycle.number,
  starting.generation, generations, mutation.rate.snp, mutation.rate.qtl)
}
\description{
Advance a family
}

