% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advance_population.R
\name{advance.population}
\alias{advance.population}
\title{Advance a population}
\usage{
advance.population(genome, list.of.families, pop.type = "random", generations,
  cycle.number, starting.generation, mutation.rate.snp, mutation.rate.qtl)
}
\description{
Takes a list of populations from a cross and advance each population either 
through random mating or through inbreeding.
}

