% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAP_data.R
\docType{data}
\name{CAP.haploids}
\alias{CAP.haploids}
\title{CAP genotypic data in haploid matrix format}
\format{A data frame with 1528 rows and 1590 columns}
\source{
\url{https://triticeaetoolbox.org/barley/}
}
\usage{
CAP.haploids
}
\description{
A \code{matrix} of genotype data in haploid format. The column
names are the names of the SNP markers and the row names are the names 
of the lines. Each line has two names with the suffix "1" or "2" to denote
the haplotype number. These haploids were created using the 
"CAP_data_preparation.R" script.
}
\keyword{datasets}

