% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAP_data.R
\docType{data}
\name{CAP.hmp}
\alias{CAP.hmp}
\title{CAP genotypic data in rrBLUP encoding}
\format{A data frame with 2832 rows and 768 columns:
\describe{
   \item{rs}{The SNP marker name}
   \item{alleles}{The SNP marker alleles}
   \item{chrom}{The chromosome in which the marker resides}
   \item{pos}{The genetic position (in thousandths of cM)}
}}
\source{
\url{https://triticeaetoolbox.org/barley/}
}
\usage{
CAP.hmp
}
\description{
A \code{data.frame} of genotype data in rrBLUP encoding. The column
names are metadata variables and line names. This genotype data is unfiltered.
}
\keyword{datasets}

