% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAP_data.R
\docType{data}
\name{CAP.lines}
\alias{CAP.lines}
\title{The names of the CAP lines}
\format{A character with 764 elements.}
\source{
\url{https://triticeaetoolbox.org/barley/}
}
\usage{
CAP.lines
}
\description{
A \code{character} of CAP line names for which genotypic data was 
used in the simulations.
}
\keyword{datasets}

