% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAP_data.R
\docType{data}
\name{CAP.markers}
\alias{CAP.markers}
\title{Marker information}
\format{A data frame with 1590 rows and 4 columns
\describe{
   \item{rs}{The SNP marker name}
   \item{alleles}{The SNP marker alleles}
   \item{chrom}{The chromosome in which the marker resides}
   \item{pos}{The genetic position (in Morgans)}
}}
\source{
\url{https://triticeaetoolbox.org/barley/}
}
\usage{
CAP.markers
}
\description{
A \code{data.frame} of marker metadata.
}
\keyword{datasets}

