% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_functions.R
\name{design.M}
\alias{design.M}
\title{M matrix}
\usage{
design.M(n)
}
\arguments{
\item{n}{The number of rows of matrix X}
}
\description{
Creates an M matrix, or the orthagonal projector
}
\references{
Rincent, R., Laloe, D., Nicolas, S., Altmann, T., Brunel, D., Revilla, P., 
Moreau, L. (2012). Maximizing the Reliability of Genomic Selection by 
Optimizing the Calibration Set of Reference Individuals: Comparison of 
Methods in Two Diverse Groups of Maize Inbreds (Zea mays L.). Genetics, 
192(2), 715–728. http://doi.org/10.1534/genetics.112.141473
}

