% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_functions.R
\name{find.pos}
\alias{find.pos}
\title{Finds the whole genome positions of QTL and markers}
\usage{
find.pos(genome, genos)
}
\arguments{
\item{genome}{The list of hypred genomes.}

\item{genos}{The n x m matrix of genotypes for n entries across m loci.}
}
\description{
Finds the whole genome positions of QTL and markers
}

