% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_loci.R
\name{genotype.loci}
\alias{genotype.loci}
\title{Genotype loci}
\usage{
genotype.loci(haploid.genos, genome, include.QTL = FALSE, error.rate = 0,
  missing.prop = 0)
}
\arguments{
\item{haploid.genos}{A 2n x m matrix of haploid alleles at m loci for n
individuals.}

\item{genome}{The hypred genome.}

\item{include.QTL}{Logical whether to include the QTL in the design matrix.
Default is \code{FALSE}.}

\item{error.rate}{The rate of genotyping errors for a particular individual.
Used to draw from a binomial distribution whether to introduce error. The 
conversion of a heterozygous genotype to one of the homozygotes or from a 
homozygous genotype to a heterozygote or one of the homozygotes will occur
will equal probability.}

\item{missing.prop}{The proportion of missing data across the genotype calls.
Genotype calls are randomly assigned missing to obtain the desired missingness
proportion.}
}
\description{
Creates a loci design matrix given a haploid matrix or a list of haploid 
matricies. This function essentially "genotypes" the individuals with
100% accuracy.
}

