% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypic_value.R
\name{genotypic.value}
\alias{genotypic.value}
\title{Calculate the genotypic value of individuals}
\usage{
genotypic.value(genome, haploid.genos)
}
\arguments{
\item{genome}{The hypred genome.}

\item{haploid.genos}{A 2n x m matrix of haploid alleles at m loci for n
individuals.}
}
\description{
Calculates the genotypic value of a set of lines based on their QTL alleles 
and the effect of those alleles.
}

