% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_functions.R
\name{invert.mat}
\alias{invert.mat}
\title{Hidden functions}
\usage{
invert.mat(x, silent = FALSE)
}
\arguments{
\item{x}{A matrix to be inverted.}

\item{silent}{Logical whether to print the method of inversion used.}
}
\description{
Attempts to find the inverse of a matrix. If unsuccessful, find the
generalized inverse.
}

