% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_crossing_block.R
\name{make.crossing.block}
\alias{make.crossing.block}
\title{Create a crossing block from known parents}
\usage{
make.crossing.block(parent1.lines, parent2.lines, n.crosses,
  method = "random", use.parents.once = FALSE)
}
\arguments{
\item{parent1.lines}{A character vector of lines for the first parents.}

\item{parent2.lines}{See \code{parent1.lines}.}

\item{n.crosses}{The number of crosses to make.}

\item{method}{The mating scheme. Can be \code{"random"} (default) or 
\code{"all.pairwise"}. If \code{"all.pairwise"}, parents will be used
more than once.}

\item{use.parents.once}{A logical whether to only use parents once.}
}
\description{
Create a crossing block based on a vector of line names for each parents 
(this can be the same vector) and a number of crosses, while choosing whether 
or not to use parents once.
}

