% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_family.R
\name{make.family}
\alias{make.family}
\title{Make a family}
\usage{
make.family(genome, parent1.genome, parent2.genome, N, generations,
  pop.type = "random", cycle.number, family.number, mutate = TRUE,
  mutation.rate.snp, mutation.rate.qtl)
}
\arguments{
\item{genome}{The list of hypred genomes.}

\item{parent1.genome}{A 2 x m matrix of the parent 1 haploid genomes.}

\item{parent2.genome}{See \code{parent1.genome}}

\item{N}{The number of F_1 individuals to generate.}

\item{generations}{The number of selfing generations of single-seed descent. 
For instance \code{generations = 5} would results in an F_6 population. 
For each generation, each of the F_1 individuals results in a single new 
individual that has undergone an additional generation of inbreeding}

\item{pop.type}{The type of population to generate. Can be \code{"inbred"} 
or \code{"random"} (default).}

\item{cycle.number}{The integer number of the breeding cycle.}

\item{family.number}{The integer number of the family.}

\item{mutate}{A logical as to generate random mutations when recombining.}

\item{mutation.rate.snp}{The per-base mutation rate of the SNPs.}

\item{mutation.rate.qtl}{The per-base mutation rate of the QTL.}
}
\description{
Make a family
}

