% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_genome.R
\name{make.genome}
\alias{make.genome}
\title{Make a genome}
\usage{
make.genome(n.chr, chr.len, n.chr.snps, genetic.map)
}
\arguments{
\item{n.chr}{The number of chromosomes in the genome}

\item{chr.len}{A vector of length \code{n.chr} with elements corresponding to
the Morgan length of each chromosome.}

\item{n.chr.snps}{A vector of length \code{n.chr} with elements corresponding
to the number of SNP loci on each chromosome.}

\item{genetic.map}{A list of length \code{n.chr} where each element is a vector
of Morgan positions corresponding to the SNP loci.}
}
\value{
A \code{list} of \code{hypred} genomes.
}
\description{
Make a genome
}

