% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_population.R
\name{make.population}
\alias{make.population}
\title{Make a population}
\usage{
make.population(genome, parental.haploids, crossing.block, N, cycle.number,
  generations, pop.type = "random", mutate = TRUE, mutation.rate.snp,
  mutation.rate.qtl)
}
\arguments{
\item{genome}{The hypred genome.}

\item{parental.haploids}{The complete 2n x m matrix of parent haploid
genotype information.}

\item{N}{The number of F_1 individuals to generate.}

\item{cycle.number}{The integer number of the breeding cycle.}

\item{generations}{The number of selfing generations of single-seed descent. 
For instance \code{generations = 5} would results in an F_6 population. 
For each generation, each of the F_1 individuals results in a single new 
individual that has undergone an additional generation of inbreeding}

\item{pop.type}{The type of population to generate. Can be \code{"inbred"} 
or \code{"random"} (default).}

\item{mutate}{A logical as to generate random mutations when recombining.}

\item{mutation.rate.snp}{The per-base mutation rate of the SNPs.}

\item{mutation.rate.qtl}{The per-base mutation rate of the QTL.}
}
\description{
Takes haploid genome information and a crossing block and produce a list of 
families. The function also requires the genome, the size of the families 
and the mutation rates.
}

