% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictions.R
\name{make.predictions}
\alias{make.predictions}
\title{Make genomic predictions}
\usage{
make.predictions(pheno.train, geno.train, geno.pred, model = "RRBLUP")
}
\arguments{
\item{pheno.train}{A matrix of phenotypes for the training population.}

\item{geno.train}{An incidence matrix of genotypes for the training population.}

\item{geno.pred}{An incidence matrix of genotypes for the prediction population
(i.e. selection candidates).}
}
\value{
A matrix of GEBVs and the solution list to the mixed model
}
\description{
Uses a RR-BLUP mixed model to predict marker effects given phenotypic and
genotypic data from a training population. Then uses the marker effects
and genotypic data from selection candidates to predict genotypic values for
those candidates. The REML method of estimating variances is used.
}

