% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_af.R
\name{measure.af}
\alias{measure.af}
\title{Calculate allele frequencies}
\usage{
measure.af(genome, haploid.genos)
}
\arguments{
\item{genome}{The hypred genome.}

\item{genos}{The 2n x m matrix of haploid genotypes for n entries and m SNPs
and QTL.}
}
\value{
A vector of allele frequencies for m loci.
}
\description{
Calculates the frequency of alleles at SNP markers or QTL. Outputs the 
minor allele frequency if requested.
}

