% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_expected_het.R
\name{measure.expected.het}
\alias{measure.expected.het}
\title{Measure expected heterozygosity}
\usage{
measure.expected.het(genos)
}
\arguments{
\item{genos}{The n x m matrix of genotypes for n entries across m loci.}
}
\description{
Calculated expected heterozygosity as the mean expected heterozygosity
across all m loci
}

