% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_LD.R
\name{measure.LD}
\alias{measure.LD}
\title{Calculate linkage disequilibrium}
\usage{
measure.LD(genome, genos, Morgan.window = NULL)
}
\arguments{
\item{genome}{The list of hypred genomes.}

\item{genos}{The n x m matrix of genotypes for n entries across m loci, 
including QTL.}

\item{Morgan.windwo}{The distance in each direction of a QTL to the markers
used to calculate LD. If \code{NULL}, all markers in the genome are used
to calculate LD with each QTL.}
}
\description{
Measures linkage disequilibrium as the correlation between QTL and markers.
}

