% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_maf.R
\name{measure.maf}
\alias{measure.maf}
\title{Calculate the minor allele frequency across a genotype matrix}
\usage{
measure.maf(genos, plot.sfs = FALSE)
}
\arguments{
\item{genos}{The n x m matrix of genotypes for n entries across m loci.}

\item{plot.sfs}{A logical whether to plot the site-frequency spectrum.}
}
\value{
A vector of minor allele frequencies for m loci.
}
\description{
Calculates the minor allele frequency. Also outputs the site-frequency
spectrum, if desired.
}

