% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{nv_df}
\alias{nv_df}
\title{Parse results to data.frame}
\usage{
nv_df(x, change)
}
\arguments{
\item{x}{A vector of values}

\item{change}{The method of updating the training population}

\item{n.iters}{The number of simulation iterations}
}
\description{
Convert a vector of values to a df during data parsing
}

