% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_subset_or_whole.R
\name{optimize.subset}
\alias{optimize.subset}
\alias{optimize.whole}
\title{Optimize PEVmean or CDmean}
\usage{
optimize.subset(method, A, n.training, phenotyped.index, unphenotyped.index,
  V_a, V_e, max.iter)

optimize.whole(method, A, n.training, phenotyped.index, unphenotyped.index, V_a,
  V_e, max.iter)
}
\arguments{
\item{A}{The n x n relationship matrix.}

\item{n.training}{The target size of the training population.}

\item{phenotyped.index}{The index of the "phenotyped" entries in the matrix 
\code{A}. This can be interpreted as the entries from which you wish to
develop an optimized training population.}

\item{unphenotyped.index}{The index of "unphenotyped" entries}

\item{The}{optimization procedure. Can be \code{"PEVmean"} or \code{"CDmean"}.}
}
\description{
Optimize the PEVmean or CDmean using a subset (for faster computation) or the
whole set of individuals.

Optimize using the whole set
}
\section{Functions}{
\itemize{
\item \code{optimize.whole}: 
}}

