% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results.R
\name{parse.results}
\alias{parse.results}
\title{Parse the results of the simulation}
\usage{
parse.results(files, filename)
}
\arguments{
\item{files}{A \code{character} vector of file paths. These files should be
\code{.RData} files.}

\item{filename}{A filepath to save the parsed, collective results}
}
\description{
This function prases the R Data files created in interations
of the simulations and extract useful data
}

