% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEVmean_CDmean.R
\name{PEVmean}
\alias{CDmean}
\alias{PEVmean}
\title{PEVmean}
\usage{
PEVmean(M, Z, A, A.inv, c, V_e, V_a)

CDmean(M, Z, A, A.inv, c, V_e, V_a)
}
\arguments{
\item{M}{Orthagonal projector. See \code{Details}}

\item{Z}{Incidence matrix}

\item{A}{Covariance matrix (i.e. additive relationship matrix)}

\item{A.inv}{The inverse of A (this is supplied to reduce computation time)}

\item{c}{Matrix of contrasts}

\item{V_e}{Error variance}

\item{V_a}{Additive genetic variance}
}
\description{
Calculates the PEV mean or the CD mean from a contrasts matrix.

CDmean
}
\section{Functions}{
\itemize{
\item \code{CDmean}: 
}}

