% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_population.R
\name{phenotype.population}
\alias{phenotype.population}
\title{Generate phenotypic values}
\usage{
phenotype.population(genome, haploid.genos, V_E, V_e,
  just.geno.values = FALSE, n.env, n.rep, run.anova = FALSE)
}
\arguments{
\item{genome}{The hypred genome.}

\item{haploid.genos}{A 2n x m matrix of haploid alleles at m loci for n
individuals.}

\item{V_E}{The variance of the distribution from which to draw environmental
effects.}

\item{V_e}{The variance of the distribution from which to draw residual
effects.}

\item{just.geno.values}{Logical whether to just return the genotypic values.}

\item{n.rep}{The number of replicates of each individual in each environment.}

\item{run.anova}{Logical whether to run an anova of the simulated phenotypes.
This is useful to verify that the target heritability is being achieved.}

\item{n.evn}{The number of environments to "phenotype" the population.}
}
\description{
Generates value measures on the population of interest. The values can be the 
genotypic values (sum of the a values at each qtl) or the phenoypic values 
(genotypic values + error)
}

