% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{qtl.eff}
\alias{qtl.eff}
\title{Retrieve QTL effects from a genome}
\usage{
qtl.eff(genome)
}
\arguments{
\item{genome}{A list of hypred genomes.}
}
\description{
Retrieve QTL effects from a genome
}

