% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombine.R
\name{recombine}
\alias{recombine}
\title{Simulate recombination}
\usage{
recombine(genome, haploid.genomeA, haploid.genomeB, mutate = TRUE,
  mutation.rate.snp, mutation.rate.qtl)
}
\arguments{
\item{genome}{The list of hypred genomes.}

\item{haploid.genomeA}{The first haploid genome.}

\item{haploid.genomeB}{The second haploid genome.}

\item{mutate}{A logical as to generate random mutations when recombining.}

\item{mutation.rate.snp}{The per-base mutation rate of the SNPs.}

\item{mutation.rate.qtl}{The per-base mutation rate of the QTL.}
}
\description{
Uses the \code{\link[hypred]{hypredRecombine}} function as a base, but will 
apply that function over the chromosomes in the genome list
}

