% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_haploids.R
\name{select.haploids}
\alias{select.haploids}
\title{Subset the haploid genotypes matrix}
\usage{
select.haploids(haploid.genos, line.names)
}
\arguments{
\item{haploid.genos}{A 2n x m matrix of haploid alleles at m loci for n
individuals.}

\item{line.names}{Individuals to subset.}
}
\description{
Subset the haploid genotypes matrix
}

