% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_population.R
\name{select.population}
\alias{select.population}
\title{Apply selection on a population}
\usage{
select.population(value.mat, sel.intensity, selection = "best",
  exclusion = NULL)
}
\arguments{
\item{value.mat}{A matrix of values (phenotypic, genotypic, or breeding) on
which to apply selection.}

\item{sel.intensity}{Either the proportion to select (i.e. from 0 to 1) or
the number of individuals to select.}

\item{selection}{The method of selection. Can be \code{"best"} (default),
\code{"worst"}, or \code{"random"}.}

\item{exlcusion}{A vector of line names to exclude when making selection.}
}
\description{
Apply selection on a population
}

