% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_values.R
\name{select.values}
\alias{select.values}
\title{Subset phenotypic or genotypic values}
\usage{
select.values(pheno.values.list, line.names)
}
\arguments{
\item{pheno.values.list}{The list returned when using the function
\code{\link{GSsim.TPUpdate}[phenotype.population]}.}

\item{line.names}{Individuals for which to extract values.}
}
\description{
Subset phenotypic or genotypic values
}

