% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_plotting.R
\name{sim.ggplot}
\alias{sim.ggplot}
\title{Plot data from a data frame summary}
\usage{
sim.ggplot(df.summary, main, ylab, xlab = "Breeding Cycle", ylim = NULL,
  xlim = c(1, n.cycles + 1), col.factors, print.plot = TRUE)
}
\arguments{
\item{df.summary}{The summary data frame produced by the \code{sim.summarize}
function.}

\item{main}{The plot title.}

\item{ylab}{The y-axis title.}

\item{xlab}{The x-axis title. Defaults to "Breeding Cycle."}

\item{ylim}{The y-axis scale range.}

\item{xlim}{The x-axis scale range. Defaults to \code{c(1, (n.cycles + 1))}.}

\item{col.factors}{The factors used in coloring the points / lines on the plot.}

\item{print.plot}{Logical. Should the plot be displayed?}
}
\description{
Plot data from a data frame summary
}

