% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_plot.R
\name{sim.plot}
\alias{sim.plot}
\title{Plot a list of results}
\usage{
sim.plot(data.list, legend.pos, just.data = FALSE, ...)
}
\arguments{
\item{data.list}{A \code{list} of parsed data for a particular statistic.}

\item{legend.pos}{The position of the legend. Accepts character inputs such as
\code{"topleft"}.}

\item{just.data}{Logical whether to return the estimates of the mean and 
confidence interval}

\item{...}{Additional arguments to pass to \code{plot}, including \code{xlim},
\code{xlab}, \code{ylim}, \code{ylab}, and \code{main}.}
}
\description{
Creates a plot of the statistic of interest over the cycles in the simulation.
The input should be a list derived from parsing the results.
}

