% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_plotting.R
\name{sim.summarize}
\alias{sim.summarize}
\title{Summarize a results data frame}
\usage{
sim.summarize(df)
}
\arguments{
\item{df}{A data frame of results that includes values of a variable
and summary metadata (i.e. the cycle, change, iteration, etc.)}
}
\description{
Summarize a results data frame
}

