% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP_optimization_algorithms.R
\name{TP.optimization.algorithms}
\alias{TP.optimization.algorithms}
\title{Optimization algorithms}
\usage{
TP.optimization.algorithms(A, phenotyped.lines, unphenotyped.lines, n.TP, V_e,
  V_a, optimization.method = c("PEVmean", "CDmean"), max.iter = 800,
  use.subset = FALSE)
}
\description{
Optimization algorithms
}

