% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_architecture.R
\name{trait.architecture}
\alias{trait.architecture}
\title{Define the trait architecture}
\usage{
trait.architecture(genome, n.QTL, qtl.index = NULL, qtl.dom.index = NULL,
  qtl.perf.index = NULL, qtl.add.eff = "normal", qtl.dom.eff = NULL)
}
\arguments{
\item{genome}{The list of hypred genomes.}

\item{n.QTL}{The number of SNPs to become QTL.}

\item{qtl.index}{A list of SNP indices per chromosome to become QTL. If NULL, 
indices are randomly sampled.}

\item{qtl.dom.index}{A list of QTL indices per chromosome to have dominance 
effects.}

\item{qtl.perf.index}{A list of QTL indices per chromosome to be perfect
markers (i.e. the observed SNP is the QTL).}

\item{qtl.add.eff}{A vector of length \code{n.QTL} of additive effects of QTL.
The vector is interpreted as such: the ith element in the vector will be the
additive effect assigned to the ith QTL. Can also be \code{"normal"} to draw 
effects from a standard normal distribution or \code{"geometric"} to draw
effects from a geometric series.}

\item{qtl.dom.eff}{A vector of length \code{n.QTL} of dominance effects of QTL.
The vector is interpreted as such: the ith element in the vector will be the
dominance effect assigned to the ith QTL.}
}
\value{
A list of hypred genomes with assigned trait architecture.
}
\description{
Adds QTL to a genome. SNPs are assigned (randomly or by the user) to become
QTL. Effects are then assigned to the QTL (from various distributions). All 
QTL have additive effects, and some, all, or none can have dominance effects.
}

