% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_predictions.R
\name{validate.predictions}
\alias{validate.predictions}
\title{Calculate realized prediction accuracy}
\usage{
validate.predictions(predicted.values, observed.values, boot.reps = NULL)
}
\arguments{
\item{predicted.values}{The predicted genotypic values (i.e. GEBVs from
genomic prediction).}

\item{observed.values}{The observed genotypic values (from using the 
\code{\link{GSsim.TPUpdate}[genotypic.value]}).}

\item{boot.reps}{The number of bootstrapping replicates when estimating
the confidence interval.}
}
\description{
Calculates the correlation between predicted genotypic values and known
genotypic values. Can also calculate a boostrapped 95% confidence interval.
}

