% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_functions.R
\name{find.pos}
\alias{find.pos}
\title{Finds the whole genome positions of QTL and markers}
\usage{
find.pos(genome)
}
\arguments{
\item{genome}{The list of hypred genomes.}
}
\description{
Finds the whole genome positions of QTL and markers
}

