% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_expected_het.R
\name{measure.expected.het}
\alias{measure.expected.het}
\title{Measure expected heterozygosity}
\usage{
measure.expected.het(genome, haploid.genos)
}
\arguments{
\item{genome}{The hypred genome}

\item{haploid.genos}{The 2n x m matrix of haploid genotypes for n entries
and m SNPs and QTL.}
}
\description{
Calculated expected heterozygosity as the mean expected heterozygosity
across all m loci
}

