% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_limit.R
\name{round.limit}
\alias{round.limit}
\title{Round limit}
\usage{
\method{round}{limit}(x, limit, option)
}
\description{
A convenience function to round a number down or up if it is outside a 
lower or upper limit
}

