% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{round.limit}
\alias{round.limit}
\title{Round with a limit}
\usage{
\method{round}{limit}(x, limit, option)
}
\arguments{
\item{x}{A vector of numbers to assess}

\item{limit}{The upper or lower limit for rounding}

\item{option}{Character vector of either \code{"upper"} or \code{"lower"}
corresponding to whether the \code{limit} is an upper limit or a lower limit.}
}
\description{
A convenience function to round a number down or up if it is outside a 
lower or upper limit
}

