% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_population.R
\name{select.population}
\alias{select.population}
\title{Apply selection on a population}
\usage{
select.population(value.mat, sel.intensity, selection = "best",
  exclusion = NULL)
}
\arguments{
\item{value.mat}{A matrix of values (phenotypic, genotypic, or breeding) on
which to apply selection.}

\item{sel.intensity}{Either the proportion to select (i.e. from 0 to 1) or
the number of individuals to select.}

\item{selection}{The method of selection. Can be \code{"best"} (default),
\code{"worst"}, \code{"random"}, or \code{"tails"}.}

\item{exlcusion}{A vector of line names to exclude when making selection.}
}
\description{
Apply selection on a population
}
\details{
The \code{"best"} method selects the top \code{sel.intensity} entries based on
the values in the \code{value.mat}, the \code{"worst"} method selects the
bottom entries, the \code{"random"} method select entries randomly, and the
\code{"tails"} method select the top \eqn{sel.intensity/2} entries
and the bottom \eqn{sel.intensity/2} entries.
}

