% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_plotting.R
\name{sim.ggplot}
\alias{sim.ggplot}
\title{Plot data from a data frame summary}
\usage{
sim.ggplot(df.summary, main, ylab, xlab = "Breeding Cycle", col.factors,
  text.y.scaling = 1.05, print.plot = TRUE)
}
\arguments{
\item{df.summary}{The summary data frame produced by the \code{sim.summarize}
function.}

\item{main}{The plot title.}

\item{ylab}{The y-axis title.}

\item{xlab}{The x-axis title. Defaults to "Breeding Cycle."}

\item{col.factors}{The factors used in coloring the points / lines on the plot.}

\item{text.y.scaling}{The scale by which to alter the maximum y value. This
new y value will be used to position text denoting the subplot in a larger
ggplot (e.g. A, B, C, etc.). Defaults to 1.05, meaning that the new y value
will be 5% higher than the max.}

\item{print.plot}{Logical. Should the plot be displayed?}
}
\description{
Plot data from a data frame summary
}

